requirements.txt
# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Multi-objective Optimization
pymoo>=0.6.0

# Financial Data
yfinance>=0.2.0

# Web Framework
Flask>=2.3.0

# AI/ML for Preference Extraction
openai>=1.0.0

# Optional: Enhanced Data Visualization (if you want to add more charts)
matplotlib>=3.5.0
plotly>=5.0.0

# Optional: Additional Scientific Computing
scikit-learn>=1.0.0

# Development and Testing (optional)
pytest>=7.0.0
jupyter>=1.0.0

# Web Server (for production deployment)
gunicorn>=20.0.0

# Environment Management
python-dotenv>=0.19.0
